feat: 결제 정보 컴포넌트 분리, 덩어리 별 주석 추가, 중복 코드 제거, 더 명확한 이름으로 수정
