+++ src/basic/components/Cart.js
@@ -1,11 +1,8 @@
 import { createElement } from '../core/createElement.js';
-import {
-  addToCart,
-  handleCartItemDelete,
-  updateCartItemQuantity
-} from '../logic/cartLogic.js';
+import { addToCart, handleCartItemDelete } from '../logic/cartLogic.js';
+import { updateCartItemQuantity } from '../logic/cartLogic.js';
 
-export const CartHeader = () => {
+export const CartTitle = () => {
   return createElement('h1', {
     className: 'text-2xl font-bold mb-4',
     textContent: '장바구니'
@@ -18,51 +15,7 @@ export const CartItems = () => {
   });
 };
 
-export const DisplayPoint = () => {
-  return createElement('span', {
-    id: 'loyalty-points',
-    className: 'text-blue-500 ml-2',
-    textContent: '(포인트: 0)'
-  });
-};
-
-export const CartTotal = () => {
-  const container = createElement('div', {
-    id: 'cart-total',
-    className: 'text-xl font-bold my-4'
-  });
-
-  const totalText = createElement('span', {
-    textContent: '총액: 0원'
-  });
-  const pointDisplay = DisplayPoint();
-
-  console.log('111', pointDisplay);
-
-  container.append(totalText, pointDisplay);
-  console.log('222', container);
-  return container;
-};
-
-export const QuantityButton = (itemId, isIncrease, label) => {
-  return createElement('button', {
-    className: 'quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1',
-    'data-product-id': itemId,
-    'data-change': isIncrease ? 1 : -1,
-    textContent: label,
-    onclick: () => updateCartItemQuantity(itemId, isIncrease)
-  });
-};
-
-export const RemoveButton = (itemId) => {
-  return createElement('button', {
-    className: 'remove-item bg-red-500 text-white px-2 py-1 rounded',
-    'data-product-id': itemId,
-    textContent: '삭제',
-    onclick: () => handleCartItemDelete(itemId)
-  });
-};
-
+//장바구니 상품
 export const CartItem = (itemToAdd) => {
   const itemSpan = createElement('span', {
     textContent: `${itemToAdd.name} - ${itemToAdd.price}원 x 1`
@@ -70,8 +23,8 @@ export const CartItem = (itemToAdd) => {
 
   const buttonContainer = createElement('div', {});
 
-  const minusButton = QuantityButton(itemToAdd.id, false, '-');
-  const plusButton = QuantityButton(itemToAdd.id, true, '+');
+  const minusButton = QuantityChangeButton(itemToAdd.id, false, '-');
+  const plusButton = QuantityChangeButton(itemToAdd.id, true, '+');
   const removeButton = RemoveButton(itemToAdd.id);
 
   buttonContainer.append(minusButton, plusButton, removeButton);
@@ -86,11 +39,33 @@ export const CartItem = (itemToAdd) => {
   return container;
 };
 
+//장바구니에 추가하는 버튼
 export const AddToCartButton = () => {
   return createElement('button', {
     id: 'add-to-cart',
     className: 'bg-blue-500 text-white px-4 py-2 rounded',
     textContent: '추가',
     onclick: () => addToCart()
+  });
+};
+
+//장바구니 상품 수량 변경 버튼
+export const QuantityChangeButton = (itemId, isIncrease, label) => {
+  return createElement('button', {
+    className: 'quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1',
+    'data-product-id': itemId,
+    'data-change': isIncrease ? 1 : -1,
+    textContent: label,
+    onclick: () => updateCartItemQuantity(itemId, isIncrease)
+  });
+};
+
+//장바구니에서 삭제하는 버튼
+export const RemoveButton = (itemId) => {
+  return createElement('button', {
+    className: 'remove-item bg-red-500 text-white px-2 py-1 rounded',
+    'data-product-id': itemId,
+    textContent: '삭제',
+    onclick: () => handleCartItemDelete(itemId)
   });
 };
