+++ src/basic/logic/stockLogic.js
@@ -1,8 +1,8 @@
 import { productData } from '../data/data.js';
 
-export const updateSelectedOptions = (selectedOptions) => {
-  const stockStatus = document.getElementById('stock-status');
-  updateStockInfo(stockStatus);
+//상품 데이터 초기화
+export const initSelectableData = (selectedOptions) => {
+  updateStockInfo();
   productData.forEach(function (item) {
     const option = document.createElement('option');
     option.value = item.id;
@@ -12,8 +12,10 @@ export const updateSelectedOptions = (selectedOptions) => {
   });
 };
 
-export const updateStockInfo = (stockInfo) => {
+//재고 상태 업데이트
+export const updateStockInfo = () => {
   let infoMessage = '';
+  const stockStatus = document.getElementById('stock-status');
   productData.forEach(function (item) {
     if (item.quantity < 5) {
       infoMessage +=
@@ -25,5 +27,5 @@ export const updateStockInfo = (stockInfo) => {
         '\n';
     }
   });
-  stockInfo.textContent = infoMessage;
+  stockStatus.textContent = infoMessage;
 };
